refactor: 클럽 행사 페이지 ClubEvents 컴포넌트 분리
