+++ src/components/ClubEvents/ClubEvent.style.ts
@@ -1,0 +1,13 @@
+import Theme from '@/styles/Theme';
+import styled from '@emotion/styled';
+
+const EmptyClubEvent = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 60vh;
+  font-size: ${Theme.fontSize.mediumContent};
+`;
+
+export { EmptyClubEvent };
