+++ src/pages/club/ClubEventPage/ClubEventPage.style.ts
@@ -1,4 +1,3 @@
-import Theme from '@/styles/Theme';
 import styled from '@emotion/styled';
 
 const ContentSpacer = styled.div`
@@ -13,19 +12,10 @@ const ContentContainer = styled.div`
   height: 100%;
 `;
 
-const EmptyClubEvent = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 60vh;
-  font-size: ${Theme.fontSize.mediumContent};
-`;
-
 const ButtonWrapper = styled.div`
   position: fixed;
   bottom: 3rem;
   right: 5rem;
 `;
 
-export { ContentContainer, ContentSpacer, ButtonWrapper, EmptyClubEvent };
+export { ContentContainer, ContentSpacer, ButtonWrapper };
